fix(schema): sửa kiểu loyalty_points và xóa @Lob trên membershipTier
loyalty_points đổi từ INT UNSIGNED sang BIGINT UNSIGNED để khớp với
kiểu Long trong entity — Hibernate validation lỗi khi khởi động do
không match giữa DB và entity mapping.

@Lob trên field membershipTier (ENUM) bị xóa vì không phù hợp — @Lob
dành cho BLOB/CLOB, dùng trên ENUM có thể gây mapping sai với một số
JPA provider.

+++ server/src/main/resources/db/migration/V1__init_schema.sql
@@ -74,7 +74,7 @@ CREATE TABLE IF NOT EXISTS `user_profiles` (
 CREATE TABLE IF NOT EXISTS `customers` (
   `id`              BIGINT UNSIGNED  NOT NULL AUTO_INCREMENT,
   `user_id`         BIGINT UNSIGNED  NOT NULL,
-  `loyalty_points`  INT UNSIGNED     NOT NULL DEFAULT 0,
+  `loyalty_points`  BIGINT UNSIGNED  NOT NULL DEFAULT 0,
   `total_spent`     DECIMAL(15,2)    NOT NULL DEFAULT 0,
   `membership_tier` ENUM('BRONZE', 'SILVER', 'GOLD', 'PLATINUM') NOT NULL DEFAULT 'BRONZE',
   `created_at`      DATETIME         NOT NULL DEFAULT CURRENT_TIMESTAMP,
